Lorem ipsum

+++ src/main/java/com/practice/librarysystem/book/BookService.java
@@ -4,7 +4,15 @@ import com.practice.librarysystem.author.Author;
 import com.practice.librarysystem.author.AuthorRepository;
 import com.practice.librarysystem.category.Category;
 import com.practice.librarysystem.category.CategoryRepository;
+import com.practice.librarysystem.exception.ForbiddenAccessException;
 import com.practice.librarysystem.exception.NotFoundException;
+import com.practice.librarysystem.statistics.user.UserAuthor;
+import com.practice.librarysystem.statistics.user.UserAuthorRepository;
+import com.practice.librarysystem.statistics.user.UserCategory;
+import com.practice.librarysystem.statistics.user.UserCategoryRepository;
+import com.practice.librarysystem.user.Role;
+import com.practice.librarysystem.user.User;
+import com.practice.librarysystem.user.UserRepository;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
@@ -12,6 +20,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @RequiredArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
@@ -21,6 +31,10 @@ public class BookService {
     BookRepository bookRepository;
     AuthorRepository authorRepository;
     CategoryRepository categoryRepository;
+    UserRepository userRepository;
+
+    UserCategoryRepository userCategoryRepository;
+    UserAuthorRepository userAuthorRepository;
 
     public Page<Book> findAllByMultipleParams(String search, Long authorId, Long categoryId,
                                               int from, int size) {
@@ -77,14 +91,21 @@ public class BookService {
         }
 
 
-        return bookRepository.findAll(pageable);
+        return bookRepository.findAllByOrderByPopularity(pageable);
     }
 
     public Book findById(Long id) {
         return findByIdOrElseThrow(id);
     }
 
-    public Book createNew(Book book, Long authorId, Long categoryId) {
+    public Book createNew(Book book, Long authorId, Long categoryId, String requesterEmail) {
+        User user = findUserByEmailOrElseThrow(requesterEmail);
+
+        if (user.getRole() != Role.ADMIN) {
+            throw new ForbiddenAccessException(
+                    String.format("Book creating failed for user %d", user.getId()));
+        }
+
         Author author = findAuthorByIdOrElseThrow(authorId);
 
         Category category = findCategoryByIdOrElseThrow(categoryId);
@@ -98,8 +119,15 @@ public class BookService {
         return book;
     }
 
-    public Book updateById(Long id, Book book, Long authorId, Long categoryId) {
+    public Book updateById(Long id, Book book, Long authorId, Long categoryId, String requesterEmail) {
         Book foundBook = findByIdOrElseThrow(id);
+
+        User user = findUserByEmailOrElseThrow(requesterEmail);
+
+        if (user.getRole() != Role.ADMIN) {
+            throw new ForbiddenAccessException(
+                    String.format("Book creating failed for user %d", user.getId()));
+        }
 
         if (book.getTitle() != null) {
             foundBook.setTitle(book.getTitle());
@@ -137,27 +165,60 @@ public class BookService {
         return foundBook;
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String requesterEmail) {
         findByIdOrElseThrow(id);
 
+        User user = findUserByEmailOrElseThrow(requesterEmail);
+
+        if (user.getRole() != Role.ADMIN) {
+            throw new ForbiddenAccessException(
+                    String.format("Book creating failed for user %d", user.getId()));
+        }
+
         bookRepository.deleteById(id);
+    }
+
+    public List<Book> findAllRecommended(String email) {
+        User currentUser = findUserByEmailOrElseThrow(email);
+
+        UserCategory userCategory = userCategoryRepository.findFirstByUserIdOrderByPopularity(currentUser.getId())
+                .orElseThrow(() -> new NotFoundException("Statistics error"));
+
+        UserAuthor userAuthor = userAuthorRepository.findFirstByUserIdOrderByPopularity(currentUser.getId())
+                .orElseThrow(() -> new NotFoundException("Statistics error"));
+
+        return bookRepository.findTop3ByAuthor_IdOrCategory_Id(
+                userAuthor.getAuthor().getId(), userCategory.getCategory().getId());
+    }
+
+    public Page<Book> findAllPopular(int from, int size) {
+        int pageNum = from / size;
+        Pageable pageable = PageRequest.of(pageNum, size);
+
+        return bookRepository.findAllByOrderByPopularity(pageable);
     }
 
     private Book findByIdOrElseThrow(Long id) {
         return bookRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException(
-                        String.format("Book with id %d not found", id)));
+                        String.format("Book with id: %d not found", id)));
     }
 
     private Author findAuthorByIdOrElseThrow(Long authorId) {
         return authorRepository.findById(authorId)
                 .orElseThrow(() -> new NotFoundException(
-                        String.format("Author with id %d not found", authorId)));
+                        String.format("Author with id: %d not found", authorId)));
     }
 
     private Category findCategoryByIdOrElseThrow(Long categoryId) {
         return categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new NotFoundException(
-                        String.format("Category with id %d not found", categoryId)));
+                        String.format("Category with id: %d not found", categoryId)));
+    }
+
+    private User findUserByEmailOrElseThrow(String email) {
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new NotFoundException(
+                        String.format("User with email: %s not found", email)));
     }
 }
